% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirScreen-package.R
\docType{package}
\name{AirScreen-package}
\alias{AirScreen}
\alias{AirScreen-package}
\title{AirScreen: Feature screening via adaptive iterative ridge}
\description{
Comprehensive tools for adaptive regularized screening on high‑dimensional data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Logic314/Air-HOLP}
}

}
\author{
Ibrahim Joudah \email{ibrahim.joudah@mq.edu.au}
Samuel Muller \email{samuel.muller@mq.edu.au}
Houying Zhu \email{houying.zhu@mq.edu.au}
}
\keyword{internal}
